+++ test/model/job/test_get_job.py
@@ -1,6 +1,7 @@
 from freezegun import freeze_time
 
-from model.job import create_job, get_job
+from model.job.job import get_job
+from model.job.create_job import create_job
 from test.model.job import BaseTestJob
